# rewardsrefund/rewardsrefund.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "rewardsrefund.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/redeemRewards/{user.id}/{redemption}": {
      "get": {
        "summary": "Redeem Rewards",
        "operationId": "RedeemRewards",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "parameters": [
          {
            "name": "user.id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "redemption",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "user.accountBalance",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "RefundService"
        ]
      }
    },
    "/api/users": {
      "post": {
        "summary": "Add a user",
        "operationId": "AddUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUserRequest"
            }
          }
        ],
        "tags": [
          "RefundService"
        ]
      }
    },
    "/api/v1/createRefund/{user.id}/{refundItemCost}": {
      "post": {
        "summary": "Create user refund",
        "operationId": "CreateUserRefund",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "parameters": [
          {
            "name": "user.id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "refundItemCost",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRefundRequest"
            }
          }
        ],
        "tags": [
          "RefundService"
        ]
      }
    },
    "/api/v1/updateBalance/{user.id}/{balance}": {
      "post": {
        "summary": "Update user balance",
        "operationId": "UpdateUserBalance",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "parameters": [
          {
            "name": "user.id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "balance",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "RefundService"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "summary": "Lists users",
        "operationId": "ListUser",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/User"
                },
                "error": {
                  "$ref": "#/definitions/runtimeStreamError"
                }
              },
              "title": "Stream result of User"
            }
          }
        },
        "tags": [
          "RefundService"
        ]
      }
    }
  },
  "definitions": {
    "AddUserRequest": {
      "type": "object",
      "title": "The request message for adding a user"
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "accountBalance": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "UserRefundRequest": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        },
        "refundItemCost": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "The request message for user refunds"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeStreamError": {
      "type": "object",
      "properties": {
        "grpc_code": {
          "type": "integer",
          "format": "int32"
        },
        "http_code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "http_status": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}